+++ service-base/src/main/java/org/eclipse/hono/service/amqp/UpstreamReceiverImpl.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016 Bosch Software Innovations GmbH.
+ * Copyright (c) 2016, 2017 Bosch Software Innovations GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -10,7 +10,7 @@
  *    Bosch Software Innovations GmbH - initial creation
  */
 
-package org.eclipse.hono.server;
+package org.eclipse.hono.service.amqp;
 
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicBoolean;